move lib.rs to utils.rs file
it's a bit cleaner to just put the utility functions in a module called utils.rs, and make all four modules public. then it's easy to import across modules without relying on a unified lib.rs file, which causes issues when trying to import from root.
